requirements.txt
streamlit
numpy
pillow
tensorflow
opencv-python
plotly
requests
pyperclip
cv2